+++ include/fileos/streamout.inl
@@ -27,9 +27,9 @@ IN THE SOFTWARE.
 
 namespace fileos {
 
-__forceinline void StreamOut::writeSizeString(char const* text, size_t size)
+__forceinline void StreamOut::writeSizeString(char const* text, uint32_t size)
 {
-    write<uint32_t>((uint32_t)size);
+    write<uint32_t>(size);
     write((void*)text, size);
 }
